% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{indpro}
\alias{indpro}
\title{Monthly U.S. Industrial Production}
\format{
An object of class \code{"xts"}.
}
\source{
Archive of the Federal Reserve Economic Data \href{https://fred.stlouisfed.org/series/INDPRO}{(FRED)}
}
\usage{
data(indpro)
}
\description{
Monthly data on the U.S. Industrial Production index (IP, index 2012=100, seasonally adjusted) collected from the
Federal Reserve Economic Data (FRED) archive. The IP has been used as MIDAS term in different contributions
(see, for instance, \insertCite{engle_ghysels_sohn_2013;textual}{rumidas}, \insertCite{conrad_lock_2015;textual}{rumidas}, and
\insertCite{amendola_candila_scognamillo_2017;textual}{rumidas}).
}
\examples{
head(indpro)
summary(indpro)
plot(indpro)
}
\references{
\insertAllCited{}
}
\keyword{datasets}
