% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{Inf_criteria}
\alias{Inf_criteria}
\title{Information Criteria}
\usage{
Inf_criteria(est)
}
\arguments{
\item{est}{The output of the estimation process.}
}
\value{
The resulting vector represents the AIC and BIC criteria.
}
\description{
Returns the Akaike and Bayesian information criteria.
}
\keyword{internal}
