% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{LF_f}
\alias{LF_f}
\title{Loss functions}
\usage{
LF_f(vol_est, vol_proxy)
}
\arguments{
\item{vol_est}{It is the estimated volatility.}

\item{vol_proxy}{It is the volatility proxy.}
}
\value{
The resulting vector represents the MSE and QLIKE averages.
}
\description{
Returns the MSE and QLIKE.
}
\keyword{internal}
