% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midas_functions.R
\name{MEM_MIDAS_lr_pred}
\alias{MEM_MIDAS_lr_pred}
\title{MEM-MIDAS long-run one-step-ahead predictions (with skewness parameter)}
\usage{
MEM_MIDAS_lr_pred(param, x, daily_ret, mv_m, K)
}
\arguments{
\item{param}{Vector of estimated values. It must be a six--dimensional vector. See the example below.}

\item{x}{Dependent variable, usually the realized volatility. It must be positive and "xts" object.}

\item{daily_ret}{Daily returns, which must be an "xts" object.}

\item{mv_m}{MIDAS variable already transformed into a matrix, through \code{\link{mv_into_mat}} function.}

\item{K}{Number of (lagged) realizations of the MIDAS variable to consider.}
}
\value{
The resulting vector is the log-likelihood value for each \eqn{i,t}.
}
\description{
Predicts the long-run term of the dependent variable, usually the realized volatility, for the MEM-MIDAS, with an asymmetric term linked to past negative returns.
}
\examples{
\donttest{
est_val<-c(alpha=0.10,beta=0.8,gamma=0.1,m=0,theta=-0.16,w2=5)
r_t<-sp500['/2010']
real<-(rv5['/2010'])^0.5		# realized volatility
mv_m<-mv_into_mat(real,diff(indpro),K=12,"monthly")
est_vol<-MEM_MIDAS_pred(est_val,real,r_t,mv_m,K=12)
head(est_vol)
}
}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{mv_into_mat}}.
}
\keyword{internal}
