% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midas_functions.R
\name{MEM_MIDAS_X_pred_no_skew}
\alias{MEM_MIDAS_X_pred_no_skew}
\title{MEM-MIDAS-X one-step-ahead predictions (no skewness parameter)}
\usage{
MEM_MIDAS_X_pred_no_skew(param, x, mv_m, K, z)
}
\arguments{
\item{param}{Vector of starting values}

\item{x}{Dependent variable, usually the realized volatility. It must be positive and "xts" object}

\item{mv_m}{MIDAS variable already transformed into a matrix, through \code{\link{mv_into_mat}} function}

\item{K}{Number of (lagged) realizations of the MIDAS variable to consider}

\item{z}{Additional daily variable which must be an "xts" object, and with the same length of x}
}
\value{
The resulting vector is the log-likelihood value for each \eqn{i,t}
}
\description{
Predicts the dependent variable, usually the realized volatility, for the MEM-MIDAS-X, with an additional X part (for instance, the VIX).
For details, see \insertCite{amendola2020doubly;textual}{rumidas} and \insertCite{engle_gallo_2006;textual}{rumidas}.
}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{mv_into_mat}}.
}
\keyword{internal}
