% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midas_functions.R
\name{MEM_pred}
\alias{MEM_pred}
\title{MEM one-step-ahead predictions (with skewness parameter)}
\usage{
MEM_pred(param, x, daily_ret)
}
\arguments{
\item{param}{Vector of starting values. It must be a three--dimensional vector. See the example below.}

\item{x}{Dependent variable, usually the realized volatility. It must be positive and "xts" object.}

\item{daily_ret}{Daily returns, which must be an "xts" object, and with the same length of x.}
}
\value{
The resulting vector is the one-step-ahead prediction for each \eqn{i,t}.
}
\description{
Predicts the dependent variable, usually the realized volatility, for the base MEM, with an asymmetric term linked to past negative returns.
For details, see \insertCite{engle_gallo_2006;textual}{rumidas}.
}
\examples{
# est_val<-c(alpha=0.10,beta=0.8,gamma=0.05)
# real<-(rv5['/2010'])^0.5		# realized volatility
# r_t<-sp500['/2010']
# head(MEM_pred(est_val,real,r_t))
}
\references{
\insertAllCited{}
}
\keyword{internal}
