% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midas_functions.R
\name{MEM_pred_no_skew}
\alias{MEM_pred_no_skew}
\title{MEM one-step-ahead predictions (no skewness parameter)}
\usage{
MEM_pred_no_skew(param, x)
}
\arguments{
\item{param}{Vector of starting values. It must be a two--dimensional vector. See the example below.}

\item{x}{Dependent variable, usually the realized volatility. It must be positive and "xts" object.}
}
\value{
The resulting vector is the log-likelihood value for each \eqn{i,t}.
}
\description{
Predicts the dependent variable, usually the realized volatility, for the base MEM.
For details, see \insertCite{engle_gallo_2006;textual}{rumidas}.
}
\examples{
# est_val<-c(alpha=0.10,beta=0.8)
# real<-(rv5['/2010'])^0.5		# realized volatility
# head(MEM_pred_no_skew(est_val,real))
}
\references{
\insertAllCited{}
}
\keyword{internal}
