% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{MEM_QMLE_sd}
\alias{MEM_QMLE_sd}
\title{Standard errors for the Quasi Maximum Likelihood estimator of the MEM-based models}
\usage{
MEM_QMLE_sd(est, x, x_pred)
}
\arguments{
\item{est}{It is the output of the maximum likelihood estimation process.}
}
\value{
The resulting vector represents the QML standard errors.
}
\description{
Obtains the standard errors for the Quasi Maximum Likelihood (QML) estimator.
}
\keyword{internal}
