% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midas_functions.R
\name{MEM_X_pred}
\alias{MEM_X_pred}
\title{MEM-X one-step-ahead predictions (with skewness parameter)}
\usage{
MEM_X_pred(param, x, daily_ret, z)
}
\arguments{
\item{param}{Vector of estimated values.}

\item{x}{Dependent variable, usually the realized volatility. It must be positive and "xts" object.}

\item{daily_ret}{Daily returns, which must be an "xts" object, and with the same length of x.}

\item{z}{Additional daily variable which must be an "xts" object, and with the same length of x.}
}
\value{
The resulting vector is the log-likelihood value for each \eqn{i,t}.
}
\description{
Predicts the dependent variable, usually the realized volatility, for the base MEM, with an asymmetric term
linked to past negative returns and an additional X part (for instance, the VIX).
For details, see \insertCite{amendola2020doubly;textual}{rumidas} and \insertCite{engle_gallo_2006;textual}{rumidas}.
}
\references{
\insertAllCited{}
}
\keyword{internal}
