% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{print.rumidas}
\alias{print.rumidas}
\title{Print method for 'rumidas' class}
\usage{
\method{print}{rumidas}(x, ...)
}
\arguments{
\item{x}{An object of class 'rumidas'.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Print method for 'rumidas' class
}
\keyword{internal}
