% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{QMLE_sd}
\alias{QMLE_sd}
\title{Standard errors for the Quasi Maximum Likelihood estimator of the GARCH-MIDAS-based models}
\usage{
QMLE_sd(est)
}
\arguments{
\item{est}{It is the output of the maximum likelihood estimation process.}
}
\value{
The resulting vector represents the QML standard errors.
}
\description{
Obtains the standard errors for the Quasi Maximum Likelihood (QML) estimator.
}
\keyword{internal}
