% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rv5}
\alias{rv5}
\title{S&P 500 realized variance at 5-minutes}
\format{
An object of class \code{"xts"}.
}
\source{
Realized library of the Oxford-Man Institute
}
\usage{
data(rv5)
}
\description{
Daily data on the realized variance of the S&P 500 collected from the realized library of
the Oxford-Man Institute \insertCite{heber_2009}{rumidas}. The realized variance has been calculated
using intradaily intervals of five minutes \insertCite{andersen_boll_1998}{rumidas}.
}
\examples{
head(rv5)
summary(rv5)
plot(rv5)
}
\references{
\insertAllCited{}
}
\keyword{datasets}
