% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sum_X_f}
\alias{sum_X_f}
\title{Summation function for the multi-step-ahead predictions of the GARCH--MIDAS models with the '--X' part.}
\usage{
sum_X_f(COEF, DELTA, H)
}
\arguments{
\item{COEF}{The sum of the parameters \eqn{alpha}, \eqn{\beta}, and \eqn{\gamma/2}, if present.}

\item{DELTA}{The AR coefficient.}

\item{H}{The length of the multi-step-ahead predictions.}
}
\value{
The vector of the summation for each H.
}
\description{
For details, see Eq. (20) of \insertCite{amendola_candila_gallo_2020;textual}{rumidas}.
}
\references{
\insertAllCited{}
}
\keyword{internal}
