% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{summary.rumidas}
\alias{summary.rumidas}
\title{Summary method for 'rumidas' class}
\usage{
\method{summary}{rumidas}(object, ...)
}
\arguments{
\item{object}{An object of class 'rumidas', that is the result of a call to \code{\link{ugmfit}} or \code{\link{umemfit}}.}

\item{...}{Additional arguments affecting the summary produced.}
}
\description{
Summary method for 'rumidas' class
}
\examples{
\donttest{
r_t<-sp500['2003/2010']
real<-(rv5['2003/2010'])^0.5		# realized volatility
fit<-umemfit(model="MEM",skew="NO",x=real)
summary.rumidas(fit)
}
}
