% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vix}
\alias{vix}
\title{VIX daily data}
\format{
An object of class \code{"xts"}.
}
\source{
\href{https://finance.yahoo.com/}{Yahoo Finance}
}
\usage{
data(vix)
}
\description{
Daily data on VIX collected from Yahoo Finance site. The VIX data
have been de-annualized and multiplied by \eqn{100^{-1}}.
}
\examples{
head(vix)
summary(vix)
plot(vix)
}
\keyword{datasets}
